requirements/base.txt
Django==1.7.1
django-model-utils==2.2
lxml==3.4.0
psycopg2==2.5.4
pyoai==2.4.4
sqlparse==0.1.13
Unipath==1.0
wsgiref==0.1.2